remove battlebot and rename concurrent battlebot to battlebot
