+++ chatgpt_allcapsbot.py
